Improve error handling for native file API.
This still seems to be a bit messy when you iniate multiple downloads. This
could be a limitation of the implementation in Chrome 78 rather than a
bug in our code though. The simple case always works, one download in one
directory where a user has allowed permissions. Multiple downloads in
differing file trees fails though.

This adds basic toasts for permissions, general write errors, and success.